Read and write the contents of usefiles properly.
    Still to do: Merge it into the current symbol table.

+++ parse/dump.c
@@ -50,7 +50,7 @@ void dumpsym(Sym *s, FILE *fd)
 
 static void outstab(Stab *st, FILE *fd, int depth)
 {
-    int i, n;
+    size_t i, n;
     void **k;
     char *ty;
